+++ tests/testthat.R
@@ -1,4 +1,4 @@
-library(testthat)
-library(dsrTest)
-
-test_check("dsrTest")
+if(require(testthat, quietly = TRUE)){
+  library(dsrTest)
+  test_check("dsrTest")
+}