+++ NEWS.md
@@ -8,3 +8,11 @@
 - Improved documentation formatting
 - Example data and vignette added
 - wmtype modifications added for Beta methods
+
+# Version 0.2.1
+
+- modifactions stop workaround now asht bug fixes are implemented
+
+# Version 0.2.3
+
+- use of testthat now "tested" (suggests != depends)